+++ Services/Hotel/HMS.Hol.Dtos/RoomTypeManager/UpdateRoomTypeDto.cs
@@ -34,10 +34,10 @@ namespace HMS.Hol.Dtos.RoomTypeManager
             set => _description = value.Trim();
         }
 
-        [Range(1, int.MaxValue, ErrorMessage = "Giá phòng/giờ phải lớn hơn 0")]
-        public int PricePerHour { get; set; }
+        [Range(1, float.MaxValue, ErrorMessage = "Giá phòng/giờ phải lớn hơn 0")]
+        public decimal PricePerHour { get; set; }
 
-        [Range(1, int.MaxValue, ErrorMessage = "Giá phòng/giờ phải lớn hơn 0")]
-        public int PricePerNight { get; set; }
+        [Range(1, float.MaxValue, ErrorMessage = "Giá phòng/giờ phải lớn hơn 0")]
+        public decimal PricePerNight { get; set; }
     }
 }
